refactor(frontend): scope generation feature module
