+++ app/frontend/src/features/generation/components/GenerationParameterForm.vue
