+++ app/frontend/src/features/generation/components/index.ts
@@ -1,0 +1,5 @@
+export { default as GenerationActiveJobsList } from './GenerationActiveJobsList.vue';
+export { default as GenerationParameterForm } from './GenerationParameterForm.vue';
+export { default as GenerationResultsGallery } from './GenerationResultsGallery.vue';
+export { default as GenerationStudio } from './GenerationStudio.vue';
+export { default as GenerationSystemStatusCard } from './GenerationSystemStatusCard.vue';
