+++ app/frontend/src/features/generation/composables/index.ts
