+++ app/frontend/src/features/generation/composables/useGenerationOrchestrator.ts
