+++ app/frontend/src/features/generation/composables/useGenerationUpdates.ts
