+++ app/frontend/src/features/generation/composables/useJobQueuePolling.ts
