+++ app/frontend/src/features/generation/composables/useJobQueueTransport.ts
