+++ app/frontend/src/features/generation/index.ts
@@ -1,0 +1,4 @@
+export * from './components';
+export * from './composables';
+export * from './services';
+export * from './stores';
