+++ app/frontend/src/features/generation/services/generationService.ts
