+++ app/frontend/src/features/generation/services/index.ts
@@ -1,4 +1,3 @@
-
 export * from './updates';
 export * from './generationService';
 export * from './validation';
