+++ app/frontend/src/features/generation/services/queueClient.ts
