+++ app/frontend/src/features/generation/services/updates.ts
