+++ app/frontend/src/features/generation/services/validation.ts
