+++ app/frontend/src/features/generation/services/websocketManager.ts
