+++ app/frontend/src/features/generation/stores/connection.ts
