+++ app/frontend/src/features/generation/stores/form.ts
