+++ app/frontend/src/features/generation/stores/index.ts
