+++ app/frontend/src/features/generation/stores/results.ts
