+++ app/frontend/src/services/index.ts
@@ -1,6 +1,6 @@
 export * from './apiClient';
 export * from './backendClient';
-export * from './generation';
+export * from '../features/generation/services';
 export * from './system';
 export * from './lora';
 export * from './history';
