+++ app/frontend/src/stores/index.ts
@@ -1,7 +1,7 @@
 export * from './settings';
 export * from './app';
 export * from './adminMetrics';
-export * from './generation';
+export * from '../features/generation/stores';
 export * from './adapterCatalog';
 export * from './performanceAnalytics';
