+++ app/frontend/src/views/DashboardView.vue
@@ -131,7 +131,7 @@ const panelConfigs = [
     placeholder:
       'Use the inline studio for quick jobs or switch to the dedicated page for the full orchestrator experience.',
     fallback: 'Loading generation studio…',
-    loader: () => import('@/components/generation/GenerationStudio.vue').then((module) => module.default),
+    loader: () => import('@/features/generation/components/GenerationStudio.vue').then((module) => module.default),
   },
   {
     key: 'gallery',
